Riddle 24: Change return type from Int to Long. (#6)
Fixes #5

+++ src/main/kotlin/com/vanniktech/rxriddles/Riddle24.kt
@@ -9,7 +9,7 @@ object Riddle24 {
    *
    * Use case: Know how many emissions have been sent out.
    */
-  fun solve(source: Observable<Any>): Single<Int> {
+  fun solve(source: Observable<Any>): Single<Long> {
     TODO()
   }
 }
